+++ web/tests/dispatch.test.js
@@ -537,6 +537,11 @@ run_test("realm settings", ({override}) => {
     assert_same(realm.realm_signup_announcements_stream_id, 41);
     realm.realm_signup_announcements_stream_id = -1; // make sure to reset for future tests
 
+    event = event_fixtures.realm__update__zulip_update_announcements_stream_id;
+    dispatch(event);
+    assert_same(realm.realm_zulip_update_announcements_stream_id, 42);
+    realm.realm_zulip_update_announcements_stream_id = -1; // make sure to reset for future tests
+
     event = event_fixtures.realm__update__default_code_block_language;
     dispatch(event);
     assert_same(realm.realm_default_code_block_language, "javascript");