+++ zerver/lib/import_realm.py
@@ -957,6 +957,9 @@ def do_import_realm(import_dir: Path, subdomain: str, processes: int = 1) -> Rea
         data, "zerver_realm", "new_stream_announcements_stream", related_table="stream"
     )
     re_map_foreign_keys(data, "zerver_realm", "signup_announcements_stream", related_table="stream")
+    re_map_foreign_keys(
+        data, "zerver_realm", "zulip_update_announcements_stream", related_table="stream"
+    )
     if "zerver_usergroup" in data:
         update_model_ids(UserGroup, data, "usergroup")
         for setting_name in Realm.REALM_PERMISSION_GROUP_SETTINGS: